git-svn-id: https://opengl-scut.googlecode.com/svn/trunk@180 81c3a017-244d-672b-1dc6-c8d6962b81c2

+++ Code/Scar/Scar/AlphaChangeUIAnimator.cpp
@@ -35,7 +35,7 @@ bool AlphaChangeUIAnimator::animateUIObject( IUIObject* node, u32 timeMS )
 	LastTime = timeMS;
 
 	//AlphaChange
-	unsigned int alpha = SrcAlpha + t * AlphaFactor;
+	unsigned int alpha = (u32)(SrcAlpha + t * AlphaFactor);
 	
 	UIObject* pNode = static_cast< UIObject* >( node );
